removed not necessary tests for ProfileImageFetchable
